+++ tailwind.config.js
@@ -1,4 +1,4 @@
 
-module.exports = {
-    plugins: [require('@tailwindcss/line-clamp')]
-}
+// module.exports = {
+//     plugins: [require('@tailwindcss/line-clamp')]
+// }
