+++ views/homepage.handlebars
@@ -1,4 +1,4 @@
-<body class="body-bg min-h-screen pt-12 md:pt-20 pb-6 px-2 md:px-0" style="font-family:'Lato',sans-serif;">
+{{!-- <body class="body-bg min-h-screen pt-12 md:pt-20 pb-6 px-2 md:px-0" style="font-family:'Lato',sans-serif;"> --}}
         {{!-- <header class="max-w-lg mx-auto">
             <a href="#">
                 <h1 class="text-4xl font-bold text-black text-center">Nobot Reviews</h1>
@@ -43,4 +43,4 @@
             <a href="#" class="hover:underline">Privacy</a>
         </footer> --}}
         <script src="../tailwind.config.js"></script>
-</body>
+{{!-- </body> --}}
