+++ views/index.html
@@ -1,4 +1,4 @@
-<!-- <!DOCTYPE html>
+<!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="UTF-8">
@@ -55,4 +55,4 @@
             <a href="#" class="hover:underline">Privacy</a>
         </footer>
 </body>
-</html> -->
+</html>
