+++ views/layouts/main.handlebars
@@ -18,7 +18,10 @@
 <body class="body-bg min-h-screen pt-12 md:pt-20 pb-6 px-2 md:px-0" style="font-family:'Lato',sans-serif;">
   <div class="flex-column justify-space-around">
     <header class="max-w-lg mx-auto">
-
+                <h1 class="text-4xl font-bold text-white text-center">Nobot Reviews</h1>
+            </a>
+        </header>
+    
       <nav>
         <!-- Conditionally render login or logout links -->
         {{#if logged_in}}
