+++ views/login.handlebars
@@ -1,4 +1,4 @@
-<body class="body-bg min-h-screen pt-12 md:pt-20 pb-6 px-2 md:px-0" style="font-family:'Lato',sans-serif;">
+{{!-- <body class="body-bg min-h-screen pt-12 md:pt-20 pb-6 px-2 md:px-0" style="font-family:'Lato',sans-serif;"> --}}
         <header class="max-w-lg mx-auto">
             <a href="#">
                 <h1 class="text-4xl font-bold text-white text-center">Nobot Reviews</h1>
@@ -39,6 +39,5 @@
             <span class="mx-3">•</span>
             <a href="#" class="hover:underline">Privacy</a>
         </footer>
-</body>
 
 <script src="./js/login.js"></script>
