+++ views/signup.handlebars
@@ -1,4 +1,4 @@
-<body class="body-bg min-h-screen pt-12 md:pt-20 pb-6 px-2 md:px-0" style="font-family:'Lato',sans-serif;">
+
     <header class="max-w-lg mx-auto">
         <a href="#">
             <h1 class="text-4xl font-bold text-white text-center">Nobot Reviews</h1>
@@ -12,7 +12,7 @@
         </section>
 
         <section class="mt-10">
-            <form class="flex flex-col login-form" method="POST" action="#">
+            <form class="flex flex-col signup-form">
                 <div class="mb-6 pt-3 rounded bg-gray-200">
                     <label class="block text-gray-700 text-sm font-bold mb-2 ml-3" for="email">User Name</label>
                     <input type="text" id="name-signup"
@@ -30,13 +30,12 @@
                 </div>
 
                 <button
-                    class="bg-purple-600 hover:bg-purple-700 text-white font-bold py-2 rounded shadow-lg hover:shadow-xl transition duration-200 signup-form"
+                    class="bg-purple-600 hover:bg-purple-700 text-white font-bold py-2 rounded shadow-lg hover:shadow-xl transition duration-200"
                     type="submit">Sign Up</button>
             </form>
         </section>
 
     </main>
 
-</body>
 
-<script src="./js/login.js"></script>
+<script src="./js/signup.js"></script>
